Moves all existing tests close to src #218
