+++ src/polyfill.js
@@ -1,3 +1,9 @@
+import present from 'present';
+
 export const system = {
   window: typeof window === 'undefined' ? global : window
+};
+
+global.performance = {
+  now: present
 };
