--- test/polyfill.js
@@ -1,5 +1,0 @@
-import present from 'present';
-
-global.performance = {
-  now: present
-};
